[chore] App: rename files
