+++ app/src/test/java/uk/ryanwong/skycatnews/domain/model/newslist/NewsTypeTest.kt
@@ -1,12 +1,11 @@
 /*
- * Copyright (c) 2022. Ryan Wong (hello@ryanwong.co.uk)
+ * Copyright (c) 2023. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.skycatnews.newslist.domain.model
+package uk.ryanwong.skycatnews.domain.model.newslist
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
-import uk.ryanwong.skycatnews.domain.model.newslist.NewsType
 
 internal class NewsTypeTest : FreeSpec() {
