+++ app/src/test/java/uk/ryanwong/skycatnews/domain/model/storydetail/StoryTest.kt
@@ -1,13 +1,11 @@
 /*
- * Copyright (c) 2022. Ryan Wong (hello@ryanwong.co.uk)
+ * Copyright (c) 2023. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.skycatnews.storydetail.domain.model
+package uk.ryanwong.skycatnews.domain.model.storydetail
 
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
-import uk.ryanwong.skycatnews.domain.model.storydetail.Content
-import uk.ryanwong.skycatnews.domain.model.storydetail.Story
 import uk.ryanwong.skycatnews.storydetail.data.local.entity.ContentEntity
 
 internal class StoryTest : FreeSpec() {
@@ -22,7 +20,7 @@ internal class StoryTest : FreeSpec() {
                 // When
                 val story = Story.fromEntity(
                     storyEntity = storyEntity,
-                    contentEntities = contentEntities
+                    contentEntities = contentEntities,
                 )
 
                 // Then
@@ -37,7 +35,7 @@ internal class StoryTest : FreeSpec() {
                 // When
                 val story = Story.fromEntity(
                     storyEntity = storyEntity,
-                    contentEntities = contentEntities
+                    contentEntities = contentEntities,
                 )
 
                 // Then
@@ -47,21 +45,21 @@ internal class StoryTest : FreeSpec() {
                     date = "2022-05-21T00:00:00Z",
                     headline = "some-headline",
                     heroImageAccessibilityText = "some-accessibility-text",
-                    heroImageUrl = "https://some.hero.image/url"
+                    heroImageUrl = "https://some.hero.image/url",
                 )
             }
 
             "Should fill headline with empty string if it comes as null" {
                 // Given
                 val storyEntity = StoryTestData.mockStoryEntity.copy(
-                    headline = null
+                    headline = null,
                 )
                 val contentEntities = listOf<ContentEntity>()
 
                 // When
                 val story = Story.fromEntity(
                     storyEntity = storyEntity,
-                    contentEntities = contentEntities
+                    contentEntities = contentEntities,
                 )
 
                 // Then
@@ -71,7 +69,7 @@ internal class StoryTest : FreeSpec() {
                     date = "2022-05-21T00:00:00Z",
                     headline = "",
                     heroImageAccessibilityText = "some-accessibility-text",
-                    heroImageUrl = "https://some.hero.image/url"
+                    heroImageUrl = "https://some.hero.image/url",
                 )
             }
 
@@ -83,7 +81,7 @@ internal class StoryTest : FreeSpec() {
                 // When
                 val story = Story.fromEntity(
                     storyEntity = storyEntity,
-                    contentEntities = contentEntities
+                    contentEntities = contentEntities,
                 )
 
                 // Then
@@ -92,12 +90,12 @@ internal class StoryTest : FreeSpec() {
                     contents = listOf(
                         Content.Paragraph(
                             text = "some-text-1",
-                        )
+                        ),
                     ),
                     date = "2022-05-21T00:00:00Z",
                     headline = "some-headline",
                     heroImageAccessibilityText = "some-accessibility-text",
-                    heroImageUrl = "https://some.hero.image/url"
+                    heroImageUrl = "https://some.hero.image/url",
                 )
             }
         }
